remove NewVmWizard
