--- src/components/Wizard/NewVmWizard/index.js
@@ -1,1 +1,0 @@
-export { NewVmWizard } from './NewVmWizard';
