+++ src/index.js
@@ -2,6 +2,5 @@ export * from './components/HelloWorld';
 export * from './components/Button';
 export * from './components/Form';
 export * from './components/Wizard/CreateVmWizard';
-export * from './components/Wizard/NewVmWizard';
 export * from './k8s/request';
 export * from './constants';
